CB#7 - Refactoring - Remove Infrastructure project
