+++ src/crypto-bank.Common/CommonProjectServiceCollectionExtensions.cs
@@ -2,9 +2,9 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace crypto_bank.Common;
 
-public static class CommonServiceCollectionExtensions
+public static class CommonProjectServiceCollectionExtensions
 {
-    public static IServiceCollection AddCommon(this IServiceCollection services)
+    public static IServiceCollection AddCommonProject(this IServiceCollection services)
     {
         services.AddSingleton<IClock, Clock>();
