--- src/crypto-bank.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -1,17 +1,0 @@
-using crypto_bank.Infrastructure.Common;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace crypto_bank.Infrastructure;
-
-public static class InfrastructureServiceCollectionExtensions
-{
-    public static IServiceCollection AddInfrastructure(
-        this IServiceCollection services,
-        ConfigurationManager configuration)
-    {
-        services.AddSingleton<IPasswordHasher, PasswordHasher>();
-
-        return services;
-    }
-}
