+++ src/crypto-bank.WebAPI/Common/Errors/Exceptions/Base/ErrorException.cs
@@ -1,0 +1,9 @@
+namespace crypto_bank.WebAPI.Common.Errors.Exceptions.Base;
+
+public abstract class ErrorException : Exception
+{
+    public ErrorException(string message, Exception? innerException = null)
+        : base(message, innerException)
+    {
+    }
+}
