+++ src/crypto-bank.WebAPI/Features/Accounts/Requests/Create.cs
@@ -2,10 +2,10 @@ using System.Security.Claims;
 using crypto_bank.Common;
 using crypto_bank.Database;
 using crypto_bank.Domain.Features.Accounts.Models;
-using crypto_bank.WebAPI.Errors;
+using crypto_bank.WebAPI.Common.Errors.Exceptions;
+using crypto_bank.WebAPI.Common.Validation;
 using crypto_bank.WebAPI.Features.Accounts.Errors;
 using crypto_bank.WebAPI.Features.Accounts.Options;
-using crypto_bank.WebAPI.Validation;
 using FluentValidation;
 using JetBrains.Annotations;
 using MediatR;
