+++ src/crypto-bank.WebAPI/Features/Auth/Requests/Authenticate.cs
@@ -1,4 +1,4 @@
-using crypto_bank.WebAPI.Validation;
+using crypto_bank.WebAPI.Common.Validation;
 using FluentValidation;
 using MediatR;
