+++ src/crypto-bank.WebAPI/Features/Users/Requests/Register.RequestHandler.cs
@@ -2,8 +2,8 @@ using crypto_bank.Common;
 using crypto_bank.Database;
 using crypto_bank.Domain.Authorization;
 using crypto_bank.Domain.Features.Users.Models;
-using crypto_bank.Infrastructure.Common;
-using crypto_bank.WebAPI.Errors;
+using crypto_bank.WebAPI.Common.Errors.Exceptions;
+using crypto_bank.WebAPI.Common.Services;
 using crypto_bank.WebAPI.Features.Users.Errors;
 using crypto_bank.WebAPI.Features.Users.Options;
 using JetBrains.Annotations;
