+++ src/crypto-bank.WebAPI/Features/Users/Requests/Register.cs
@@ -1,5 +1,5 @@
 using crypto_bank.Common;
-using crypto_bank.WebAPI.Validation;
+using crypto_bank.WebAPI.Common.Validation;
 using FluentValidation;
 using MediatR;
