+++ src/crypto-bank.WebAPI/crypto-bank.WebAPI.csproj
@@ -9,6 +9,7 @@
 
   <ItemGroup>
     <PackageReference Include="FluentValidation" Version="11.5.2" />
+    <PackageReference Include="Isopoh.Cryptography.Argon2" Version="1.1.12" />
     <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
     <PackageReference Include="MediatR" Version="12.0.1" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.5" />
@@ -19,7 +20,6 @@
   <ItemGroup>
     <ProjectReference Include="..\crypto-bank.Common\crypto-bank.Common.csproj" />
     <ProjectReference Include="..\crypto-bank.Database\crypto-bank.Database.csproj" />
-    <ProjectReference Include="..\crypto-bank.Infrastructure\crypto-bank.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
